+++ src/parseIngredient.ts
@@ -5,10 +5,10 @@ import {
   forsRegEx,
   ofRegEx,
   rangeSeparatorRegEx,
+  trailingQuantityRegEx,
   unitsOfMeasure,
 } from './constants';
 import type { Ingredient, ParseIngredientOptions, UnitOfMeasure } from './types';
-import { compactStringArray } from './utils';
 
 const newLineRegExp = /\r?\n/;
 
@@ -18,7 +18,9 @@ const addIdToUomDefinition = ([uom, def]: [string, UnitOfMeasure]) => ({ id: uom
  * Parses a string into an array of recipe ingredient objects
  */
 export const parseIngredient = (
-  /** The ingredient text. */
+  /**
+   * The ingredient list, as plain text.
+   */
   ingredientText: string,
   /**
    * Configuration options. Defaults to {@link defaultOptions}.
@@ -30,7 +32,12 @@ export const parseIngredient = (
   const uomArray = Object.entries(mergedUOMs).map(addIdToUomDefinition);
   const uomArrayLength = uomArray.length;
 
-  return compactStringArray(ingredientText.split(newLineRegExp)).map(line => {
+  const ingredientArray = ingredientText
+    .split(newLineRegExp)
+    .map(line => line.trim())
+    .filter(Boolean);
+
+  return ingredientArray.map(line => {
     const oIng: Ingredient = {
       quantity: null,
       quantity2: null,
@@ -41,15 +48,55 @@ export const parseIngredient = (
     };
 
     // Check if the first character is numeric.
-    const nqResultFirstChar = numericQuantity(line.substring(0, 1));
+    if (isNaN(numericQuantity(line[0]))) {
+      // The first character is not numeric. First check for trailing quantity/uom.
+      const trailingQtyResult = trailingQuantityRegEx.exec(line);
 
-    if (isNaN(nqResultFirstChar)) {
-      // The first character is not numeric, so the entire line is the description.
-      oIng.description = line;
+      if (trailingQtyResult) {
+        // Trailing quantity detected.
+        // Remove the quantity and unit of measure from the description.
+        oIng.description = line.replace(trailingQuantityRegEx, '').trim();
 
-      // If the line ends with ":" or starts with "For ", then it is assumed to be a group header.
-      if (oIng.description.endsWith(':') || forsRegEx.test(oIng.description)) {
-        oIng.isGroupHeader = true;
+        // Trailing quantity/range.
+        const firstQty = trailingQtyResult[3];
+        const secondQty = trailingQtyResult[12];
+        if (!firstQty) {
+          oIng.quantity = numericQuantity(secondQty);
+        } else {
+          oIng.quantity = numericQuantity(firstQty);
+          oIng.quantity2 = numericQuantity(secondQty);
+        }
+
+        // Trailing unit of measure.
+        const uomRaw = trailingQtyResult.at(-1);
+        if (uomRaw) {
+          let uom = '';
+          let uomID = '';
+          let i = -1;
+
+          while (++i < uomArrayLength && !uom) {
+            const { alternates, id, short, plural } = uomArray[i];
+            const versions = [...alternates, id, short, plural];
+            if (versions.includes(uomRaw)) {
+              uom = uomRaw;
+              uomID = id;
+            }
+          }
+
+          if (uom) {
+            oIng.unitOfMeasureID = uomID;
+            oIng.unitOfMeasure = opts.normalizeUOM ? uomID : uom;
+          }
+        }
+      } else {
+        // The first character is not numeric, and no trailing quantity was detected,
+        // so the entire line is the description.
+        oIng.description = line;
+
+        // If the line ends with ":" or starts with "For ", then it is assumed to be a group header.
+        if (oIng.description.endsWith(':') || forsRegEx.test(oIng.description)) {
+          oIng.isGroupHeader = true;
+        }
       }
     } else {
       // The first character is numeric. See how many of the first seven